Adapt RSL to new indicator types
